Rebasing (#10)
* Built theAutoSteer module (#64)

Built the AutoSteer module.

This module is responsible for converting desired and actual altitude and heading values and turning them into the desired roll, pitch, and rudder angles as well as desired airspeed required to achieve the above heading and altitude.

As of this commit, the implementations are very simplistic. This module has only been designed to function at cruising speed and will not properly takeoff or land. This commit has in mind our first autonomy goal. Past that, this module will have to be revisited.

* Replaced all ASSERT_ commands to EXPECT_ to make testing easier. In the case that a test fails, we will know all of the checks that failed instead of just the first check that failed.

* Telemetry (#76)

Created Telemetry state machine structure, added Mavlink Lib2

* Simulation (#72)

* some changes

* some minor updates

* Finished the infrastructure for sending things to flightgear (though it still needs to be tuned)

* got rid of a build directory

* modified cmakelists to accomodate infinion

* Added final touches to support hooking into Simulation. The other modifications are there to support the Infinion airframe

* Moved the stuff that sends Simulation data to flight gear into the Simulink-Sim repo rather than having it here

* Rather than cluttering up our production code with #ifdefs, I created an Intercept folder that contains source files that can be pulled in instead of some hardware specific files in Autopilot during Simulation builds

* Safety Code (#77)

* Ok, PPM catches the interrupt now.

But the biggest changes are to the way I propose we do things in Safety. The way I've set it up, the files that get generated by cubeMX get pulled in by the build system, no more of this copy and paste to other random ass files nonsense. That also means that you can generate whenever and things will still build. And if you want to modify a pin mapping or something, just update the cubeMX file and hit generate, and thats it. The other thing is the code we write will live exclusively in /Src and /Inc. I could see no reason to have the library folder. Now the build system is rigged to not do shit with the stuff in library. Last thing to note is that the main file gets generated by cubeMX (so lives in boardFiles/Src), but we can modify it in specific sections without the next code generation session overwriting those sections.

* CubeMX unfortunately only generates a main.c file. So for it to update properly, you have to rename main.cpp to main.c before generation, then back to main.cpp after generation. I've just updated the cmake script to fail if it finds a main.c file, to prevent some confusion

* Nurse PPM manager back in

* remove all the stuff we don't use

* PWM and Safety Controller addition (#68)

* PWM Manager Works

Note: VERIFY TIMER SETTINGS!!

* Add in safety controller

* Fixed the math... I think?

* Fixed timing fr

* Address Changes

Co-authored-by: dakaleyt <32375512+dakaleyt@users.noreply.github.com>

* Safety beginnings ppm (#73)

* PPM changes

using timer to count the gaps,
converting the gaps using the max/min
get_us and get written
not tested

* added the PPM width stuff

might have to move the pulse width stuff to before converting it to us, if accuracy is bad

* Added in Pulse Width and Interchip code

- added pulse width to PPM.cpp
- added in interchip code from safety-nuclear-changes

Co-authored-by: Aaditya <bluebird1339@gmail.com>

* Complete Main Function

* Actual code that Infinion flew with on  first ZeroPilot flight ever (#75)

Co-authored-by: Anthony Berbari <Anthony.berbari@uwaterloo.ca>

* remove IWDG

Co-authored-by: Anthony Berbari <fberbari@edu.uwaterloo.ca>
Co-authored-by: Anthony Berbari <32182266+Fberbari@users.noreply.github.com>
Co-authored-by: dakaleyt <32375512+dakaleyt@users.noreply.github.com>
Co-authored-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>
Co-authored-by: Aaditya <bluebird1339@gmail.com>
Co-authored-by: Anthony Berbari <Anthony.berbari@uwaterloo.ca>

* Fixing hal shitshow (#78)

* Ok things are fixed now, though I'll wait for the safety branch merges before merging this

* Fixed the entire Hal layer

Now, all our code will reference the drivers that are directly generated by CubeMX.

* deleted the old library files

* Revert "Telemetry (#76)" (#79)

This reverts commit 8db384a35db1b96e35ef2600448829944a5ea2f4.

* Rebasing (#9)

* Commit Boardfile Changes (#82)

* Attitude manager cleanup (#81)

* Checks NAN case for MadgwickAHRSupdate

* Gets yaw percent from GetFromPathManager.cpp

* Created a new state called fetchSensorMeasurementsMode that is run before executing SensorFusionMode. The new state gets the sensor data and the SensorFusionMode uses the collected data instead of having direct access to sensor objects.

* Create a state for the attitude manager so we know when it has completed a cycle

* Unit tests converted to accommodate new state

* Telemetry Manager Fixed Changes (#84)

- telemetry manager state machine and tests

Co-authored-by: Sahil Kale <32375512+sahil-kale@users.noreply.github.com>
Co-authored-by: NixonCTChan <68513632+NixonCTChan@users.noreply.github.com>

Co-authored-by: Anthony Berbari <32182266+Fberbari@users.noreply.github.com>
Co-authored-by: JTing86 <jingting9786@gmail.com>
Co-authored-by: Sahil Kale <32375512+sahil-kale@users.noreply.github.com>
Co-authored-by: Anthony Berbari <fberbari@edu.uwaterloo.ca>
Co-authored-by: dakaleyt <32375512+dakaleyt@users.noreply.github.com>
Co-authored-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>
Co-authored-by: Aaditya <bluebird1339@gmail.com>
Co-authored-by: Anthony Berbari <Anthony.berbari@uwaterloo.ca>
Co-authored-by: NixonCTChan <68513632+NixonCTChan@users.noreply.github.com>